phpstan: liga parseModelCastsMethod e zera os 223 erros (99% eram falso positivo)
Causa raiz: o Larastan so entende casts declarados via `protected $casts`
(estilo antigo), nao via metodo `casts(): array` (Laravel 11+, usado em
todo o projeto) — sem a flag parseModelCastsMethod, toda propriedade
com cast (datas, enums, decimais) resolvia como string/mixed em
qualquer arquivo de app/, gerando ~220 dos 223 erros.

Depois de ligar a flag, sobraram 99 erros reais pra corrigir:
- imovel_tipo no PDF do contrato mostrava o valor bruto do enum
  ('apartment') em vez do label em portugues ('Apartamento') —
  bug de conteudo real, nao so de tipagem.
- foreach de fotos de ocorrencia podia quebrar se o campo viesse como
  um unico UploadedFile em vez de array (Arr::wrap).
- alguns array_map/metodos sem tipo de parametro/retorno documentado.
- abort_unless($model, ...) trocado por abort_unless($model !== null, ...)
  em varios controllers (o helper do Laravel nao tem $boolean tipado
  de verdade, mas o phpdoc do Larastan assume que tem).
- ~45 nullsafe (?->) redundantes em relacoes/colunas que o schema
  garante nao-nulas, simplificados para -> direto.

3 casos ficaram como falso positivo confirmado (nao "corrigidos" —
suprimidos com comentario explicando o porque): Seeder::$command
realmente pode ser null fora do artisan (bati nisso via tinker nesta
sessao), o metodo routeNotificationFor() em canais de notificacao
customizados nao tem interface publica no Laravel, e o config do
Media Library referencia a classe do pacote Pro que este projeto nao
instala de proposito.

Suite de testes rodada com sqlite em memoria (workaround pro bug de
I/O do bind mount no Windows) confirma zero regressao: mesma
contagem de falhas pre-existentes (o bug de sessao/CSRF 419 ja
identificado antes, e um MP_ACCESS_TOKEN nao carregado nos testes —
nenhum dos dois toca o codigo alterado aqui).

+++ app/Services/MercadoPagoService.php
@@ -513,7 +513,7 @@ class MercadoPagoService
             throw new \InvalidArgumentException('State invalido (assinatura nao confere).');
         }
 
-        if (now()->timestamp - $timestamp > self::STATE_TTL_SECONDS) {
+        if ((int) now()->timestamp - $timestamp > self::STATE_TTL_SECONDS) {
             throw new \InvalidArgumentException('Link de conexao expirado. Gere um novo.');
         }
 
@@ -600,8 +600,12 @@ class MercadoPagoService
                 );
             }
 
-            $expiresAt = $receiver->mp_token_expires_at?->timestamp ?? 0;
-            $isExpiringSoon = $expiresAt - now()->timestamp < self::TOKEN_REFRESH_MARGIN_SECONDS;
+            // phpstan claims mp_token_expires_at is never null here, but the
+            // column is genuinely nullable (a receiver can have an access
+            // token without ever having had an expiry recorded) — keep the
+            // nullsafe.
+            $expiresAt = (int) ($receiver->mp_token_expires_at?->timestamp ?? 0); // @phpstan-ignore nullsafe.neverNull
+            $isExpiringSoon = $expiresAt - (int) now()->timestamp < self::TOKEN_REFRESH_MARGIN_SECONDS;
 
             if (! $isExpiringSoon) {
                 return $this->assertOrdersApiAccessToken($receiver->mp_access_token);
@@ -651,6 +655,7 @@ class MercadoPagoService
     }
 
     /**
+     * @param  array<string, mixed>  $body
      * @return array{access_token: string, refresh_token?: string, expires_in: int, user_id: int, live_mode?: bool}
      */
     private function requestToken(array $body): array
